Add ResourceManagers(Texture, PSO, RenderPass, Shader) Add ShaderGroupClass(For Pso, Material) NormalMapping(DX11 Only)

+++ Daydream/Source/Daydream/Graphics/API/OpenGL/OpenGLTextureCube.h
@@ -10,7 +10,7 @@ namespace Daydream
 		// Constrcuter Destructer
 		OpenGLTextureCube(const TextureDesc& _desc, const Array<Array<UInt8>>& _initialData);
 		virtual ~OpenGLTextureCube();
-		virtual inline void* GetNativeHandle() override { return reinterpret_cast<void*>(textureID); }
+		virtual inline void* GetNativeHandle() override { return reinterpret_cast<void*>(static_cast<UInt64>(textureID)); }
 	private:
 		UInt32 textureID;
 	};
